Fix example in Chrome
When `connectedCallback` runs for `<auto-complete>` element while the
document is still parsing, its contents is still unavailable.
https://developer.mozilla.org/en-US/docs/Web/Web_Components/Using_custom_elements#Using_the_lifecycle_callbacks

It is not clear whether we need to restructure our `connectedCallback`
to be more resilient to this case. In the meantime, fix the example
document by defining the `auto-complete` element at a time that the
document is fully loaded.