Replace `sr-update` event dispatch with a MutationObserver
The event fired has no effect in production which indicates that it
shouldn't be in the code at all. What we want to do in the tests is to
wait for the feedback element to have text so we should instead use a
MutationObserver to wait for the element to change before running our
assertion.