Pull request suggestions (#54)
* Add missing `package-lock.json` file

* Inline all the screen reader announcements

By inlining the screen reader announcement text we remove a layer of
indirection.

* Replace `sr-update` event dispatch with a MutationObserver

The event fired has no effect in production which indicates that it
shouldn't be in the code at all. What we want to do in the tests is to
wait for the feedback element to have text so we should instead use a
MutationObserver to wait for the element to change before running our
assertion.

* Inline listbox ID in tests

We remove a level of indirection by inlining the list box ID. It's more
apparent to the reader of the tests what elements are being referred to
without having to find the definition of the `listboxId` variable.